Time: 55 ms (90.42%), Space: 33.2 MB (33.46%) - LeetHub

+++ 1603-design-parking-system/1603-design-parking-system.cpp
@@ -1,6 +1,6 @@
 class ParkingSystem {
 private:
-    int big, medium, small;
+    short big, medium, small;
 public:
     ParkingSystem(int big, int medium, int small) {
         this->big = big;
